% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdp_plot_sig_uncertainty.R
\name{hdp_plot_sig_uncertainty}
\alias{hdp_plot_sig_uncertainty}
\title{Plots point estimates and credible intervals of mutational}
\usage{
hdp_plot_sig_uncertainty(Phi, dst = NULL)
}
\arguments{
\item{Phi}{A 3D array of dimension V x K x S, where
V is the number of categories, K is the number of clusters (signatures) and
S is the number of posterior samples}

\item{dst}{The path of a directory where the signatures will be stored the allocation variables of N observations.}
}
\description{
Plots point estimates and credible intervals of mutational
}
