% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdpExtra-package.R
\docType{package}
\name{hdpExtra-package}
\alias{hdpExtra}
\alias{hdpExtra-package}
\title{hdpExtra: An uncertainty-aware extension of the hdp package for mutational signatures}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
