% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdpExtra_posterior.R
\name{hdpExtra_posterior}
\alias{hdpExtra_posterior}
\title{An extended version of hdp::hdp_posterior}
\usage{
hdpExtra_posterior(
  hdp,
  burnin,
  n,
  space,
  cpiter = 1,
  seed = sample(1:10^7, 1),
  verbosity = 0
)
}
\arguments{
\item{hdp}{A hdpState object}

\item{burnin}{The number of burn-in iterations.}

\item{n}{The number of posterior samples to collect.}

\item{space}{The number of iterations between collected samples.}

\item{cpiter}{The number of iterations of concentration parameter sampling to perform after each iteration.}

\item{seed}{The (integer) seed that can be set to reproduce output. Default is a
random seed from 1 -- 10^7, reported in the output.}

\item{verbosity}{Verbosity of debugging statements.
0 (least verbose) -- 4 (most verbose). 0 highly recommended - only change for debugging small examples.}
}
\description{
An extended version of hdp::hdp_posterior implementing the Gibbs
sampler described in Teh et al 2006. Our version provides draws from the
distribution over allocation variables and signature parameters.
}
