% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdpExtra_clases.R
\docType{class}
\name{HdpExtraChain-class}
\alias{HdpExtraChain-class}
\title{An S4 class to store an MCMC chain}
\description{
An S4 class to store an MCMC chain
}
\section{Slots}{

\describe{
\item{\code{hdpChain}}{An hdpSampleChain object for integration with the hdp package}

\item{\code{allocations}}{A matrix where each column is a partition of the
observations sampled from the posterior distribution}

\item{\code{Phi}}{A list whose elements are draws from the posterior over signature parameters}
}}

