% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdpExtra_clases.R
\docType{class}
\name{HdpExtraChainMulti-class}
\alias{HdpExtraChainMulti-class}
\title{An S4 class to store multiple MCMC chains}
\description{
An S4 class to store multiple MCMC chains
}
\section{Slots}{

\describe{
\item{\code{hdpChains}}{An hdpSampleMulti object for integration with the hdp package}

\item{\code{allocations}}{A matrix where each column is a partition of the
observations sampled from the posterior distribution}

\item{\code{Phi}}{A list whose elements are draws from the posterior over signature parameters}
}}

