% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdpExtra_clases.R
\name{HdpExtraChainMulti}
\alias{HdpExtraChainMulti}
\title{Creates an object of class HdpExtraChainMulti}
\usage{
HdpExtraChainMulti(hdpExtraChains)
}
\arguments{
\item{hdpExtraChains}{A list of objects of class HdpExtraChain}
}
\description{
Creates an object of class HdpExtraChainMulti
}
