% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdp_plot_sig_uncertainty.R
\name{hdp_plot_sig_uncertainty}
\alias{hdp_plot_sig_uncertainty}
\title{Plots point estimates and credible intervals of mutational}
\usage{
hdp_plot_sig_uncertainty(Phi, dst = NULL)
}
\arguments{
\item{dst}{The path of a directory where the signatures will be stored the allocation variables of N observations.}

\item{chains}{An object of class hdpSampleMulti}

\item{Z}{A vector of length N containing}
}
\description{
Plots point estimates and credible intervals of mutational
}
