% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdp_postprocessing.R
\name{hdp_postprocessing}
\alias{hdp_postprocessing}
\title{allocations is an N x S matrix, (N = number of mutations, S = number of posterior iterations)}
\usage{
hdp_postprocessing(hdp_extra_chains, allocations_best)
}
\arguments{
\item{hdp_extra_chains}{An object of class HdpExtraChainMulti}

\item{allocations_best}{A vector of length N containing a representative partition}
}
\value{
Point estimates of the parameters characterising a representative partition
}
\description{
allocations is an N x S matrix, (N = number of mutations, S = number of posterior iterations)
}
