% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdp_postprocessing.R
\name{hdp_postprocessing}
\alias{hdp_postprocessing}
\title{allocations is an N x S matrix, (N = number of mutations, S = number of posterior iterations)}
\usage{
hdp_postprocessing(allocations, Phi, allocations_best, alpha = 0.95)
}
\arguments{
\item{allocations}{An N x S matrix}

\item{Phi}{A list of S matrices. Each matrix of dimension V x K^(s)}

\item{allocations_best}{A vector of length N containing a representative partition}

\item{alpha}{The probability coverage of the intervals (by default 0.95)}
}
\value{
Point estimates of the parameters characterising a representative partition
}
\description{
allocations is an N x S matrix, (N = number of mutations, S = number of posterior iterations)
}
